+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 * New `read_scanimage_metadata` for reading scanimage metadata from a file directly as a python dict [#405](https://github.com/catalystneuro/roiextractors/pull/401)
+* Added volumetric imaging support with `is_volumetric` flag, `get_frame_shape`, `get_num_planes`, and `get_video_shape` methods [PR #TBD](https://github.com/catalystneuro/roiextractors/pull/XXX)
 
 
 ### Fixes